effect_001: decompile func_801A1EBC, the mesh renderer
The last function of the overlay. Each mesh part is a command stream: vertex
groups posed through one joint each via the GTE light matrix, then a triangle
list and a quad list. Every primitive is transformed twice -- with the view
matrix for the screen coordinates and depth of both the environment-mapped
shaded prim and the textured prim, then with the light matrix, whose winding
is the back-face cull when the render asks for it -- and rejected when a
projected vertex leaves the clip rectangle.

Two spellings are load-bearing. The loop body works on per-part copies of the
vertex buffer and the two prim cursors, which is what puts the long-lived
originals on the stack the way the original build has them. The clip test is
four separate ifs each storing reject = 1: cross-jumping folds their tails
into one store, so the code is the same, but the constant's reference count
is what gives it the register the original allocated.

Types settled along the way: EffectMesh is the { skeleton, parts } header
the battle slot points at, replacing the EffectSkeleton ** view; BattleGfx::ot
is the word array it is, so the two battle prototypes that take it and the
spark renderer's cast follow; the slot gains the textured RGB and part mask,
and EffectRender the two fields the function copies out.

Six PsyQ INLINE_C.H macros libgte.h did not carry go into a new
psxsdk/inline_c.h that mirrors the SDK header.

+++ include/battle/bc_object15.h
@@ -16,6 +16,6 @@ void func_800C96E4(SVECTOR *pos, s32 scale, s16 angle);
  * @brief Emit @p prim into @p ot.
  * @return The new head of the prim buffer @p head came from.
  */
-void *func_800C9E10(void *prim, u8 *ot, s32 mode, void *head);
+void *func_800C9E10(void *prim, u32 *ot, s32 mode, void *head);
 
 #endif /* BATTLE_BC_OBJECT15_H */
