+++ include/battle/bc_object16.h
@@ -17,7 +17,7 @@
  *
  * @return The new head of the list @p a3 came from.
  */
-extern s32 func_800CBC68(BattleEffectPrim *prim, u8 *ot, s32 mode, s32 head);
+extern s32 func_800CBC68(BattleEffectPrim *prim, u32 *ot, s32 mode, s32 head);
 
 /** @brief Per-frame step installed on the particle tasks. */
 extern void func_800CE158(void);
